requirements.txt
# Optimized requirements for faster Coolify deployment
# Removed heavy ML dependencies that cause long build times

# Core web framework
streamlit==1.28.0

# API clients
openai==1.61.0
mistralai==1.9.10

# Environment and HTTP
python-dotenv==1.0.0
requests==2.32.3

# Data processing - optimized for Python 3.11
numpy==1.26.4
pandas==2.2.3
pyarrow  # Added back - cmake now available in Dockerfile

# Document processing - compatible with Python 3.11
pymupdf==1.24.14
pillow==10.4.0  # Updated to stable version
docling==2.54.0

# Database
psycopg2-binary==2.9.10  # Updated to available version - should work in Docker environment

# Text processing
tiktoken==0.5.0

# Machine Learning
scikit-learn==1.3.2  # Required for cosine_similarity in 5-chat.py

# GPU acceleration for Docling (optional)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
sentence-transformers>=2.2.0

# Utilities
python-dateutil==2.8.2
beautifulsoup4==4.12.3
lxml==4.9.3  # Added back - should work with cmake available

# opencv-python-headless==4.8.0.76  # ~200MB - uncommented for Docker deployment

# Authentication
bcrypt==4.1.2